Bug corrections and minor functionality changes
The lose of one column issue with 'virtualedit' wasn't related with the
longest line as previously thought, but with the first line (smallest
one) of selection, so changing it fixes now for real (hopefully) the bug
of losing one column after moving.

Using "range" in functions that didn't used visual selection was a
problem when (a:firstline + count) greater than the last line so that
range was deleted and replaced with <C-u> at the function calls.

The name 'are_same_line' was really referring to a column, so it was
renamed after 'are_same_cols'

Create function to initialize configurations variables to its defaults.