Read v:count before it is overwritten
Fixes #42.

The v:count parameter is overwritten whenever we invoke a normal-mode
command so we must ensure that we read the value of v:count at the start
of our functions, before we call any normal-mode commands.

Another change in this commit is that the computations that use v:count
now use it as a multiplier, instead of adding it to the computed
distance. The behavior should be the same as before for MoveLineDown(1)
but it is different for larger steps such as MoveLineHalfPageDown().
In the previous version, moving half a page twice would only move it
half a page + 2 lines.