Large vertical block moves now move until the edge
If we vertically move a block using a large step when we are close to
the start or the end of the file, move as much as possible (that is,
until the first or the last line).

This should make the behavior of block moves work the same as line
moves. Previously, a block move that attempted to move out of bounds
would not move at all.