Don't use range functions for vertical line moves
Previously, if we attempted to make a large line move that would put the
line out of bounds it would usually give error E16 (Invalid Address). It
would only work if out out of bounds destination was only a single line
out before/after the start/end of the file.

Now, moving a line vertically by a large step does not produce an error.
If the step is too large then the movement is clipped to the start or
end of the file.