Clarify the move_past_end_of_line logic
Hopefully it should be more self explanatory now. The situation that we
have to worry about is when we create a selection that is already past
the end of line. For example if we have

   ABC
   DEFGH

and ask to move the rectangle "C  FGH" to the right, we want to stay
put, because we area already past the end of line. We should be careful
to not accidentally move it backwards. Without that `last < shortest`
test it would be transformedi into something like

   C  AB
   FGHED