Further consolidate the MoveVertically functions
By taking advantage of '[ and '] we can use the same method to reindent
the lines in MoveLineVertically and MoveBlockVertically. This lets us
put all the line-moving logic in a single place.

There is also a small improvement to the behavior of MoveBlockVertically
when "nomodifiable" is set. Previously we would end up in normal mode
after the function was called and now we stay in visual mode, with the
current selection preserved.