Improve the virtualedit workaround.
I reported that bug with the virtualedit to the upstream Vim maintainers
and it has since been fixed in both Vim and Neovim. However, since we
will need to still support Vim <= 8.2 for a while, I also improved the
workaround that we use.

It turns out that virtualedit is only bugged in the cases where we don't
actually need to use it. Therefore, instead of fixing the cursor after
it ends up in the wrong place, we can instead disable virtualedit when
it is not needed. One advantage of doing it this way is that the `[ and
`] marks now always point to the right place after the operation.