Stop treating linewise visual mode as a special case.
This commit removes the special case for linewise visual mode in the
ModeBlockHorizontally function. I think that it is more consistent if
all the visual modes behave the same: now it always switches to visual
mode and then tries to move the selected rectangular area.

That said, most of the time trying to move a linewise selection isn't
super useful. It only does something interesting if
move_past_end_of_line is enabled.