Improve behavior with tabs.
In this commit we replace all our uses of virtcol() by col(), which
fixes a couple of bugs related to tabs.

One bug was in move_past_end_of_line. Since strwidth treats tabs as
having 1 character, if the line had tabs we would think that the end of
line appeared sooner than it should. If we tried to move something to
the right it would get "stuck" before reaching the end.

Another bug was that MoveRight needed a count of at least 8 to be able
to move over a tab. In the fixed version a count of 1 is enough, which
makes th MoveRight binding behave more similarly to "l".

Finally, there was the minor issue that "normal! 0|" is actually an
invalid command. It does the right thing because 0 moves to the first
column and then the | is ignored but relying on a command being silently
ignored is a bit fishy. Anyway, now that we don't call `|` at all this
problem is avoided entirely.