Use silent to avoid 'Hit ENTER to continue' prompts
Some commands such as :move and >> can output messages and when there
are 2 or more lines of these messages it can cause a 'Hit ENTER to
continue' prompt, which interrupts the flow of moving things up and
down.

Instead of trying to identify exactly which chatty commands need to be
silenced, we simply use silent in the root function call, which ensures
that absolutely everything gets silenced.